requirements.txt
gradio==4.44.0
pdfplumber==0.11.4
python-docx==1.1.2
